prep for monorepo by moving `types`, `testsupport`, `client` under `auth`
These packages will otherwise collide with existing packages in
`github.com/weberc2/comments`.
