+++ pkg/auth/authservice_test.go
@@ -10,8 +10,8 @@ import (
 	"time"
 
 	"github.com/dgrijalva/jwt-go"
-	"github.com/weberc2/auth/pkg/testsupport"
-	"github.com/weberc2/auth/pkg/types"
+	"github.com/weberc2/auth/pkg/auth/testsupport"
+	"github.com/weberc2/auth/pkg/auth/types"
 	"golang.org/x/crypto/bcrypt"
 )
