+++ pkg/auth/client/auth_test.go
