+++ pkg/pgtokenstore/pgtokenstore_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/weberc2/auth/pkg/types"
+	"github.com/weberc2/auth/pkg/auth/types"
 )
 
 func TestPGTokenStore_DeleteExpired(t *testing.T) {
